--- CG_Project/3d_models/Skydome3D/Skydome.mtl
@@ -1,16 +1,0 @@
-# 3ds Max Wavefront OBJ Exporter v0.97b - (c)2007 guruware
-# File Created: 01.07.2019 22:25:14
-
-newmtl skybox_Mat
-	Ns 9.999999
-	Ni 1.500000
-	d 1.000000
-	Tr 0.000000
-	Tf 1.000000 1.000000 1.000000 
-	illum 2
-	Ka 0.588235 0.588235 0.588235
-	Kd 0.588235 0.588235 0.588235
-	Ks 0.000000 0.000000 0.000000
-	Ke 0.588235 0.588235 0.588235
-	map_Ka .\Skydome.png
-	map_Kd .\Skydome.png
