+++ CG_Project/src/CG_Project.cpp
@@ -1,7 +1,10 @@
 // Assignment 3, Part 1 and 2
 //
 // Modify this file according to the lab instructions.
-// test
+//test
+
+#include "utils.h"
+#include "utils2.h"
 
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
@@ -13,52 +16,245 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
-#include "Camera.h"
-#include "Terrain.h"
-#include <Model.h>
-#include <Skybox.h>
 #include <algorithm>
 #include <cstdlib>
-#include <glm/gtc/type_ptr.hpp>
 #include <iostream>
 
-#include <Context.h>
-#include <IOEvent.h>
-#include <utils.h>
-#include <utils2.h>
+#include <glm/gtc/type_ptr.hpp>
+
+// The attribute locations we will use in the vertex shader
+enum AttributeLocation { POSITION = 0, NORMAL = 1 };
+
+// Struct for representing an indexed triangle mesh
+struct Mesh {
+  std::vector<glm::vec3> vertices;
+  std::vector<glm::vec3> normals;
+  std::vector<uint32_t> indices;
+};
+
+// Struct for representing a vertex array object (VAO) created from a
+// mesh. Used for rendering.
+struct MeshVAO {
+  GLuint vao;
+  GLuint vertexVBO;
+  GLuint normalVBO;
+  GLuint indexVBO;
+  int numVertices;
+  int numIndices;
+};
+
+// Struct for resources and state
+struct Context {
+  int width;
+  int height;
+
+  float aspect;
+  GLFWwindow *window;
+  GLuint program;
+  Trackball trackball;
+  Mesh mesh;
+  MeshVAO meshVAO;
+  GLuint defaultVAO;
+  GLuint cubemap;
+  float elapsed_time;
+
+  // Rendering -- Assignment3
+  glm::vec4 lightColor = glm::vec4{1.0f, 1.0f, 1.0f, 1.0f};
+  glm::vec3 lightPosition = glm::vec3{1.0, 1.0, 1.0};
+  bool enableLight = true;
+
+  // float ambientColor[3] = {0.02, 0.0, 0.0};
+  glm::vec3 ambientColor = glm::vec3(0.02, 0.0, 0.0);
+  bool enableAmbiant = true;
+  glm::vec3 diffuseColor = glm::vec3{0.4, 0.0, 0.0};
+  bool enableDiffuse = true;
+  glm::vec3 specularColor = glm::vec3(0.04, 0.04, 0.04);
+  bool enableSpecular = true;
+  float specularPower = 60.0;
+
+  glm::vec4 backgroudColor = glm::vec4(0.2, 0.2, 0.2, 1.0);
+
+  bool enablePerspective = false;
+  bool ortho = true;
+  float zoomFactorOrtho = 0.15;
+  float zoomFactorPerspective = 5.0f;
+  float zoom = 0.f;
+  float fov = 50.0f;
+};
+
+// Returns the value of an environment variable
+std::string getEnvVar(const std::string &name) {
+  char const *value = std::getenv(name.c_str());
+  if (value == nullptr) {
+    return std::string();
+  } else {
+    return std::string(value);
+  }
+}
+
+// Returns the absolute path to the shader directory
+std::string shaderDir(void) {
+  std::string rootDir = getEnvVar("ASSIGNMENT3_ROOT");
+  if (rootDir.empty()) {
+    std::cout << "Error: ASSIGNMENT3_ROOT is not set." << std::endl;
+    std::exit(EXIT_FAILURE);
+  }
+  return rootDir + "/CG_Project/src/shaders/";
+}
+
+// Returns the absolute path to the 3D model directory
+std::string modelDir(void) {
+  std::string rootDir = getEnvVar("ASSIGNMENT3_ROOT");
+  if (rootDir.empty()) {
+    std::cout << "Error: ASSIGNMENT3_ROOT is not set." << std::endl;
+    std::exit(EXIT_FAILURE);
+  }
+  return rootDir + "/CG_Project/3d_models/";
+}
+
+// Returns the absolute path to the cubemap texture directory
+std::string cubemapDir(void) {
+  std::string rootDir = getEnvVar("ASSIGNMENT3_ROOT");
+  if (rootDir.empty()) {
+    std::cout << "Error: ASSIGNMENT3_ROOT is not set." << std::endl;
+    std::exit(EXIT_FAILURE);
+  }
+  return rootDir + "/CG_Project/cubemaps/";
+}
+
+void loadMesh(const std::string &filename, Mesh *mesh) {
+  OBJMesh obj_mesh;
+  objMeshLoad(obj_mesh, filename);
+  mesh->vertices = obj_mesh.vertices;
+  mesh->normals = obj_mesh.normals;
+  mesh->indices = obj_mesh.indices;
+}
+
+void createMeshVAO(Context &ctx, const Mesh &mesh, MeshVAO *meshVAO) {
+  // Generates and populates a VBO for the vertices
+  glGenBuffers(1, &(meshVAO->vertexVBO));
+  glBindBuffer(GL_ARRAY_BUFFER, meshVAO->vertexVBO);
+  auto verticesNBytes = mesh.vertices.size() * sizeof(mesh.vertices[0]);
+  glBufferData(GL_ARRAY_BUFFER, verticesNBytes, mesh.vertices.data(),
+               GL_STATIC_DRAW);
+
+  // Generates and populates a VBO for the vertex normals
+  glGenBuffers(1, &(meshVAO->normalVBO));
+  glBindBuffer(GL_ARRAY_BUFFER, meshVAO->normalVBO);
+  auto normalsNBytes = mesh.normals.size() * sizeof(mesh.normals[0]);
+  glBufferData(GL_ARRAY_BUFFER, normalsNBytes, mesh.normals.data(),
+               GL_STATIC_DRAW);
+
+  // Generates and populates a VBO for the element indices
+  glGenBuffers(1, &(meshVAO->indexVBO));
+  glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, meshVAO->indexVBO);
+  auto indicesNBytes = mesh.indices.size() * sizeof(mesh.indices[0]);
+  glBufferData(GL_ELEMENT_ARRAY_BUFFER, indicesNBytes, mesh.indices.data(),
+               GL_STATIC_DRAW);
+
+  // Creates a vertex array object (VAO) for drawing the mesh
+  glGenVertexArrays(1, &(meshVAO->vao));
+  glBindVertexArray(meshVAO->vao);
+  glBindBuffer(GL_ARRAY_BUFFER, meshVAO->vertexVBO);
+  glEnableVertexAttribArray(POSITION);
+  glVertexAttribPointer(POSITION, 3, GL_FLOAT, GL_FALSE, 0, nullptr);
+  glBindBuffer(GL_ARRAY_BUFFER, meshVAO->normalVBO);
+  glEnableVertexAttribArray(NORMAL);
+  glVertexAttribPointer(NORMAL, 3, GL_FLOAT, GL_FALSE, 0, nullptr);
+  glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, meshVAO->indexVBO);
+  glBindVertexArray(ctx.defaultVAO); // unbinds the VAO
+
+  // Additional information required by draw calls
+  meshVAO->numVertices = mesh.vertices.size();
+  meshVAO->numIndices = mesh.indices.size();
+}
+
+void initializeTrackball(Context &ctx) {
+  double radius = double(std::min(ctx.width, ctx.height)) / 2.0;
+  ctx.trackball.radius = radius;
+  glm::vec2 center = glm::vec2(ctx.width, ctx.height) / 2.0f;
+  ctx.trackball.center = center;
+}
 
 void init(Context &ctx) {
+  ctx.program =
+      loadShaderProgram(shaderDir() + "mesh.vert", shaderDir() + "mesh.frag");
 
-  /*-------terrain -------*/
-  Terrain *terrain = new Terrain;
-  terrain->load(modelDir() + "red-sculpted-plains-terrain/source/"
-                             "RedSculptedPlains Mesh Output/RedSculptedPlains "
-                             "Mesh Output 2.obj");
-  Shader *shader =
-      new Shader(shaderDir() + "mesh.vert", shaderDir() + "mesh.frag");
-  terrain->setShader(shader);
-  terrain->scale(glm::vec3(300.f, 300.f, 300.f));
-  terrain->rotate(-90.f, glm::vec3(1.f, 0.f, 0.f));
+  loadMesh((modelDir() + "gargo.obj"), &ctx.mesh);
+  createMeshVAO(ctx, ctx.mesh, &ctx.meshVAO);
 
-  ctx.terrain = terrain;
+  // Load cubemap texture(s)
+  ctx.cubemap = loadCubemap(cubemapDir() + "/Forrest/");
 
-  /*-------camera -------*/
-  ctx.camera =
-      new Camera(glm::vec3(0.0f, 0.0f, 3.0f), glm::vec3(0.0f, 0.0f, -1.0f),
-                 glm::vec3(0.0f, 1.0f, 0.0f));
+  initializeTrackball(ctx);
+}
 
-  ctx.skydome = new Model;
-  ctx.skydome->load(modelDir() + "Skydome3D/Skydome.obj");
-  ctx.skydome->scale(glm::vec3(2.f, 2.f, 2.f));
-  ctx.skydome->setShader(shader);
+// MODIFY THIS FUNCTION
+void drawMesh(Context &ctx, GLuint program, const MeshVAO &meshVAO) {
+  // Define uniforms
+  glm::mat4 model = trackballGetRotationMatrix(ctx.trackball);
+  // glm::mat4 view = glm::mat4();
+  glm::mat4 view = glm::lookAt(
+      glm::vec3(1.0f, 0.2f, 0.5f), // position of the camera
+      glm::vec3(0.0f),             // and looks at origin
+      glm::vec3(0.0f, 1.0f, 0.0f)  // normalized vector, how camera is oriented
+      );
 
-  /*ctx.skybox = new Skybox(cubemapDir() + "/Forest/");
-  ctx.skybox->setShader(
-      new Shader(shaderDir() + "skybox.vert", shaderDir() + "skybox.frag"));*/
+  glm::mat4 projection;
 
-  ctx.projection =
-      glm::perspective(glm::radians(ctx.fov),
-                       (float)ctx.width / (float)ctx.height, 0.1f, 1000.0f);
+  if (!ctx.enablePerspective) {
+    projection = glm::ortho(-ctx.aspect - ctx.zoom, ctx.aspect + ctx.zoom,
+                            -1.0f - ctx.zoom, 1.0f + ctx.zoom, 0.01f, 100.0f);
+  } else {
+
+    if (ctx.ortho) {
+      ctx.ortho = false;
+      ctx.zoom = 0;
+      ctx.fov = 50.0f;
+    }
+
+    projection =
+
+        glm::perspective(glm::radians(ctx.fov),
+                         (float)ctx.width / (float)ctx.height, 0.1f, 100.0f);
+  }
+
+  glm::mat4 mv = view * model;
+  glm::mat4 mvp = projection * mv;
+
+  // Activate program
+  glUseProgram(ctx.program);
+
+  // Bind textures
+  glActiveTexture(GL_TEXTURE1);
+  glBindTexture(GL_TEXTURE_CUBE_MAP, 1);
+
+  // Pass uniforms
+  glUniform1i(glGetUniformLocation(ctx.program, "u_cubemap"), 1);
+
+  glUniformMatrix4fv(glGetUniformLocation(program, "u_mv"), 1, GL_FALSE,
+                     &mv[0][0]);
+  glUniformMatrix4fv(glGetUniformLocation(program, "u_mvp"), 1, GL_FALSE,
+                     &mvp[0][0]);
+  glUniform1f(glGetUniformLocation(program, "u_time"), ctx.elapsed_time);
+  glUniform3fv(glGetUniformLocation(ctx.program, "u_light_position"), 1,
+               &ctx.lightPosition[0]);
+  glUniform3fv(glGetUniformLocation(ctx.program, "u_light_color"), 1,
+               ctx.enableLight ? &ctx.lightColor[0] : &glm::vec3(0.0f)[0]);
+  glUniform3fv(glGetUniformLocation(ctx.program, "u_ambient_color"), 1,
+               ctx.enableAmbiant ? &ctx.ambientColor[0] : &glm::vec3(0.0f)[0]);
+  glUniform3fv(glGetUniformLocation(ctx.program, "u_diffuse_color"), 1,
+               ctx.enableDiffuse ? &ctx.diffuseColor[0] : &glm::vec3(0.0f)[0]);
+  glUniform3fv(glGetUniformLocation(ctx.program, "u_specular_color"), 1,
+               ctx.enableSpecular ? &ctx.specularColor[0]
+                                  : &glm::vec3(0.0f)[0]);
+  glUniform1f(glGetUniformLocation(ctx.program, "u_specular_power"),
+              ctx.specularPower);
+
+  // Draw!
+  glBindVertexArray(meshVAO.vao);
+  glDrawElements(GL_TRIANGLES, meshVAO.numIndices, GL_UNSIGNED_INT, 0);
+  glBindVertexArray(ctx.defaultVAO);
 }
 
 void display(Context &ctx) {
@@ -67,9 +263,108 @@ void display(Context &ctx) {
   glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
   glEnable(GL_DEPTH_TEST); // ensures that polygons overlap correctly
-  ctx.terrain->Render(ctx);
-  // ctx.skybox->Render(ctx);
-  ctx.skydome->Render(ctx);
+  drawMesh(ctx, ctx.program, ctx.meshVAO);
+}
+
+void reloadShaders(Context *ctx) {
+  glDeleteProgram(ctx->program);
+  ctx->program =
+      loadShaderProgram(shaderDir() + "mesh.vert", shaderDir() + "mesh.frag");
+}
+
+void mouseButtonPressed(Context *ctx, int button, int x, int y) {
+  if (button == GLFW_MOUSE_BUTTON_LEFT) {
+    ctx->trackball.center = glm::vec2(x, y);
+    trackballStartTracking(ctx->trackball, glm::vec2(x, y));
+  }
+}
+
+void mouseButtonReleased(Context *ctx, int button, int x, int y) {
+  if (button == GLFW_MOUSE_BUTTON_LEFT) {
+    trackballStopTracking(ctx->trackball);
+  }
+}
+
+void moveTrackball(Context *ctx, int x, int y) {
+  if (ctx->trackball.tracking) {
+    trackballMove(ctx->trackball, glm::vec2(x, y));
+  }
+}
+
+void errorCallback(int /*error*/, const char *description) {
+  std::cerr << description << std::endl;
+}
+
+void keyCallback(GLFWwindow *window, int key, int scancode, int action,
+                 int mods) {
+  // Forward event to GUI
+  ImGui_ImplGlfwGL3_KeyCallback(window, key, scancode, action, mods);
+  if (ImGui::GetIO().WantCaptureKeyboard) {
+    return;
+  } // Skip other handling
+
+  Context *ctx = static_cast<Context *>(glfwGetWindowUserPointer(window));
+  if (key == GLFW_KEY_R && action == GLFW_PRESS) {
+    reloadShaders(ctx);
+  }
+}
+
+void charCallback(GLFWwindow *window, unsigned int codepoint) {
+  // Forward event to GUI
+  ImGui_ImplGlfwGL3_CharCallback(window, codepoint);
+  if (ImGui::GetIO().WantTextInput) {
+    return;
+  } // Skip other handling
+}
+
+void mouseButtonCallback(GLFWwindow *window, int button, int action, int mods) {
+  // Forward event to GUI
+  ImGui_ImplGlfwGL3_MouseButtonCallback(window, button, action, mods);
+  if (ImGui::GetIO().WantCaptureMouse) {
+    return;
+  } // Skip other handling
+
+  double x, y;
+  glfwGetCursorPos(window, &x, &y);
+
+  Context *ctx = static_cast<Context *>(glfwGetWindowUserPointer(window));
+  if (action == GLFW_PRESS) {
+    mouseButtonPressed(ctx, button, x, y);
+  } else {
+    mouseButtonReleased(ctx, button, x, y);
+  }
+}
+
+void cursorPosCallback(GLFWwindow *window, double x, double y) {
+  if (ImGui::GetIO().WantCaptureMouse) {
+    return;
+  } // Skip other handling
+
+  Context *ctx = static_cast<Context *>(glfwGetWindowUserPointer(window));
+  moveTrackball(ctx, x, y);
+}
+
+void resizeCallback(GLFWwindow *window, int width, int height) {
+  Context *ctx = static_cast<Context *>(glfwGetWindowUserPointer(window));
+  ctx->width = width;
+  ctx->height = height;
+  ctx->aspect = float(width) / float(height);
+  ctx->trackball.radius = double(std::min(width, height)) / 2.0;
+  ctx->trackball.center = glm::vec2(width, height) / 2.0f;
+  glViewport(0, 0, width, height);
+}
+
+void scrollCallback(GLFWwindow *window, double x, double y) {
+
+  Context *ctx = static_cast<Context *>(glfwGetWindowUserPointer(window));
+
+  if (ctx->enablePerspective) {
+
+    ctx->fov += (float)ctx->zoomFactorPerspective * y;
+
+  } else {
+    ctx->zoom += (float)ctx->zoomFactorOrtho * y;
+  }
 }
 
 void drawImGuiControls(Context &ctx) {
