--- CG_Project/src/Camera.cpp
@@ -1,86 +1,0 @@
-#include "Camera.h"
-
-Camera::Camera(const glm::vec3 &position, const glm::vec3 &lookAtPosition,
-               const glm::vec3 &upvec)
-    : Front(lookAtPosition), Up(upvec), MovementSpeed(SPEED),
-      MouseSensitivity(SENSITIVITY), Zoom(ZOOM) {
-
-  Position = position;
-  WorldUp = upvec;
-  Yaw = YAW;
-  Pitch = PITCH;
-  updateCameraVectors();
-}
-
-void Camera::setPosition(const glm::vec3 &position) { Position = position; }
-
-void Camera::setLookAtPosition(const glm::vec3 &lookat) { Front = lookat; }
-
-void Camera::setUpVector(const glm::vec3 &upvec) { Up = upvec; }
-
-glm::mat4 Camera::getCameraViewMatrix() {
-  return glm::lookAt(Position, Position + Front, Up);
-}
-
-// Processes input received from any keyboard-like input system. Accepts input
-// parameter in the form of camera defined ENUM (to abstract it from windowing
-// systems)
-void Camera::ProcessKeyboard(Camera_Movement direction, float deltaTime) {
-  float velocity = MovementSpeed * deltaTime;
-  if (direction == FORWARD)
-    Position += Front * velocity;
-  if (direction == BACKWARD)
-    Position -= Front * velocity;
-  if (direction == LEFT)
-    Position -= Right * velocity;
-  if (direction == RIGHT)
-    Position += Right * velocity;
-}
-
-// Processes input received from a mouse input system. Expects the offset value
-// in both the x and y direction.
-void Camera::ProcessMouseMovement(float xoffset, float yoffset,
-                                  bool constrainPitch) {
-  xoffset *= MouseSensitivity;
-  yoffset *= MouseSensitivity;
-
-  Yaw += xoffset;
-  Pitch += yoffset;
-
-  // Make sure that when pitch is out of bounds, screen doesn't get flipped
-  if (constrainPitch) {
-    if (Pitch > 89.0f)
-      Pitch = 89.0f;
-    if (Pitch < -89.0f)
-      Pitch = -89.0f;
-  }
-
-  // Update Front, Right and Up Vectors using the updated Euler angles
-  updateCameraVectors();
-}
-
-// Processes input received from a mouse scroll-wheel event. Only requires input
-// on the vertical wheel-axis
-void Camera::ProcessMouseScroll(float yoffset) {
-
-  Position = glm::vec3(Position.x + 1.5 * yoffset, Position.y + 1.5 * yoffset,
-                       Position.z + 1.5 * yoffset);
-}
-
-// Calculates the front vector from the Camera's (updated) Euler Angles
-void Camera::updateCameraVectors() {
-  // Calculate the new Front vector
-  glm::vec3 front;
-  front.x = cos(glm::radians(Yaw)) * cos(glm::radians(Pitch));
-  front.y = sin(glm::radians(Pitch));
-  front.z = sin(glm::radians(Yaw)) * cos(glm::radians(Pitch));
-  Front = glm::normalize(front);
-  // Also re-calculate the Right and Up vector
-  Right = glm::normalize(glm::cross(Front, WorldUp)); // Normalize the vectors,
-                                                      // because their length
-                                                      // gets closer to 0 the
-                                                      // more you look up or
-                                                      // down which results in
-                                                      // slower movement.
-  Up = glm::normalize(glm::cross(Right, Front));
-}
