--- CG_Project/src/SkyDome.h
@@ -1,13 +1,0 @@
-#ifndef SKYDOME_H
-#define SKYDOME_H
-
-#include <Model.h>
-
-class SkyDome : public Model {
-public:
-  SkyDome();
-
-  void load(const std::string &modelPath);
-};
-
-#endif // SKYDOME_H
