--- CG_Project/src/Terrain.cpp
@@ -1,5 +1,0 @@
-#include "Terrain.h"
-
-struct Context;
-
-Terrain::Terrain() : Model() {}
