--- CG_Project/src/Terrain.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <Model.h>
-
-class Terrain : public Model {
-
-public:
-  Terrain();
-};
