--- CG_Project/src/shaders/skybox.frag
@@ -1,13 +1,0 @@
-// Skybox fragment shader
-#version 150
-
-in vec3 TexCoords;
-
-out vec4 FragColor;
-
-uniform samplerCube skybox;
-
-void main()
-{
-    FragColor = texture(skybox, TexCoords);
-}
